==> modeling/3.1.feature_exclude.ipynb <==
import pandas as pd
import json
import os

def remove_excluded_features(features_data_filename="moex_with_features.csv",
                             exclude_list_filename="feature_exclude_list.json"):
    """
    Загружает CSV с фичами и JSON со списком фич для исключения,
    удаляет указанные колонки и сохраняет обновлённый CSV обратно.
    """

    # 1. Загрузка данных
    df = pd.read_csv(features_data_filename)
    print(f"""
        {len(df.columns) = } \n
        {df.columns} \n
    """
    )

    with open(exclude_list_filename, "r", encoding="utf-8") as f:
        exclude_features = json.load(f)

    # Убедимся, что список — именно список строк
    if not isinstance(exclude_features, list):
        raise ValueError("JSON должен содержать список фич для исключения")

    # 2. Удаляем колонки, которых нет — без ошибок
    df = df.drop(columns=[col for col in exclude_features if col in df.columns], errors='ignore')

    return df



data_folder = "../data/"
feature_file = os.path.join(data_folder, 'moex_with_features.csv')

config_folder = "config/"
exclude_list_filename = os.path.join(config_folder, "feature_exclude_list.json")

df = remove_excluded_features(features_data_filename=feature_file, exclude_list_filename=exclude_list_filename)

df.sort_values(by=['Date', 'Ticker'], inplace=True)

df.to_csv(feature_file, index=False)
print(f"✅ Обновлённый файл сохранён: {feature_file}")


print(f"""
    {len(df.columns) = } \n
    {df.columns} \n
"""
)
