==> modeling/4.stock_exclude.ipynb <==
import pandas as pd
import json
import os

def remove_excluded_tickers(features_data_filename="moex_with_features.csv",
                            tickers_exclude_filename="tickers_exclude_list.json"):
    """
    Загружает CSV и JSON со списком тикеров для исключения,
    удаляет строки, где колонка 'Ticker' входит в список, и сохраняет результат.
    """

    # 1. Загрузка данных
    df = pd.read_csv(features_data_filename)
    with open(tickers_exclude_filename, "r", encoding="utf-8") as f:
        exclude_tickers = json.load(f)

    # Проверка формата JSON
    if not isinstance(exclude_tickers, list):
        raise ValueError("JSON должен содержать список тикеров для исключения")

    # 2. Проверим, есть ли колонка Ticker
    if "Ticker" not in df.columns:
        raise KeyError("В файле фичей нет колонки 'Ticker'")

    # 3. Фильтруем
    initial_count = len(df)
    df = df[~df["Ticker"].isin(exclude_tickers)]
    removed_count = initial_count - len(df)

    print(f"✅ Удалено {removed_count} строк с тикерами из списка.")
    return df




data_folder = "../data/"
feature_file = os.path.join(data_folder, 'moex_with_features.csv')

config_folder = "config/"
tickers_exclude_filename = os.path.join(config_folder, "stock_exclude_list.json")

df = remove_excluded_tickers(features_data_filename=feature_file, tickers_exclude_filename=tickers_exclude_filename)
df.to_csv(feature_file, index=False)
print(f"✅ Обновлённый файл сохранён: {feature_file}")
