==> modeling/8.4.model_training_TSMixer.ipynb <==
# TSMixer moddeling

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

# from sklearn.model_selection import TimeSeriesSplit
# from sklearn.ensemble import RandomForestClassifier
# from sklearn.metrics import accuracy_score, classification_report, roc_auc_score
# from lightgbm import LGBMClassifier



import joblib
import json
import os
from datetime import datetime

from TSMixer_Classifier_modeling import select_feature_columns
from TSMixer_Classifier_modeling import walk_forward_train_TSMixer_model
from TSMixer_Classifier_modeling import train_model_TSMixer

model_name = "TSMixer_tbm10d" # ИМЯ МОДЕЛИ


# ==============================================================================
# ОСНОВНОЙ СКРИПТ МОДЕЛИРОВАНИЯ
# ==============================================================================

# --- ШАГ 1: Подготовка данных ---


data_folder = "../data/"
final_dataset_filename = 'moex_final_dataset.csv'
TARGET_COLUMN = 'tbm_10d' # Наша цель для предсказания!

# Загружаем финальный датасет
print(f"Загрузка данных из: {data_folder + final_dataset_filename}")
try:
    df = pd.read_csv(data_folder + final_dataset_filename)
    # if features_to_remove:
    #     df = df.drop(columns=features_to_remove)
    #     print(f"Удалено {len(features_to_remove)} неважных признаков.")


    df['Date'] = pd.to_datetime(df['Date'])
    df.sort_values(by='Date', inplace=True)
    df.reset_index(drop=True, inplace=True)
    print("Данные успешно загружены.")
except FileNotFoundError:
    print("ОШИБКА: Финальный датасет не найден.")
    raise


print(f"\nИсходный размер датасета: {len(df)} строк.")
df.dropna(subset=[TARGET_COLUMN], inplace=True)
df.reset_index(drop=True, inplace=True)
print(f"Размер датасета после фильтрации по CUSUM-событиям: {len(df)} строк.")


# Определяем признаки (X) и цель (y)
# Исключаем все не-признаки: идентификаторы, дату и ВСЕ колонки с метками
# tbm_cols = [col for col in df.columns if col.startswith('tbm_')]
# cols_to_drop = ['Date', 'Ticker', 'Open', 'High', 'Low', 'Close', 'Volume'] + tbm_cols

# feature_cols = [col for col in df.columns if col not in cols_to_drop]
feature_cols = select_feature_columns(df)

X = df[feature_cols]
y = df[TARGET_COLUMN].astype(int)





print(f"\nЦелевая переменная (y): {TARGET_COLUMN}")
print(f"Количество признаков (X): {len(feature_cols)}")
print("Признаки:", feature_cols)

walk_forward_train_TSMixer_model(df, X, y, lookback=32)



# # --- ШАГ 4: Анализ важности признаков ---

print("\n--- Определение важности признаков ---")
print("Переобучаем модель на всех доступных данных для получения наиболее стабильной оценки...")

# # Обучаем финальную модель на всех данных
final_model = train_model_TSMixer(X, y, epochs=10, lookback=32)


# ToDo Пока не понятно как считать важность признаков для нейросетей
# # Создаем DataFrame с важностью признаков
# feature_importance = pd.DataFrame({
#     'feature': feature_cols,
#     'importance': final_model.feature_importances_ # метод - Mean Decrease in Impurity (Gini Importance)
# }).sort_values('importance', ascending=False)

# print("Топ-20 самых важных признаков:")
# print(feature_importance.head(20))

# # Визуализируем важность
# plt.figure(figsize=(12, 8))
# sns.barplot(x='importance', y='feature', data=feature_importance.head(20), palette='viridis')
# plt.title('Топ-20 самых важных признаков для модели', fontsize=16)
# plt.xlabel('Важность', fontsize=12)
# plt.ylabel('Признак', fontsize=12)
# plt.tight_layout()
# plt.show()


# print("Топ-20 самых НЕ важных признаков:")
# print(feature_importance.tail(20))
# # Визуализируем важность
# plt.figure(figsize=(12, 8))
# sns.barplot(x='importance', y='feature', data=feature_importance.tail(20), palette='viridis')
# plt.title('Топ-20 самых НЕ важных признаков для модели', fontsize=16)
# plt.xlabel('Важность', fontsize=12)
# plt.ylabel('Признак', fontsize=12)
# plt.tight_layout()
# plt.show()


# --- ШАГ 5: Сохранение артефактов модели ---

print("\n--- Сохранение артефактов модели ---")

output_folder = "../models/"

# timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
# filename_base = f"{model_name}_{timestamp}"
filename_base = model_name

# Создаем папку, если она не существует
os.makedirs(output_folder, exist_ok=True)
print(f"Артефакты будут сохранены в папку: {output_folder}")

# 1. Сохранение самой модели
model_path = os.path.join(output_folder, f"{filename_base}.joblib")
joblib.dump(final_model, model_path)
print(f"Модель сохранена в: {model_path}")

# 2. Сохранение списка признаков
features_path = os.path.join(output_folder, f"{filename_base}_features.json")
features_list = list(X.columns)
with open(features_path, 'w') as f:
    json.dump(features_list, f, indent=4)
print(f"Список признаков сохранен в: {features_path}")

# 3. Сохранение метаданных и результатов
metadata_path = os.path.join(output_folder, f"{filename_base}_metadata.json")

# Собираем все важные параметры и результаты в один словарь
metadata = {
    "model_name": model_name,
    "model_class": type(final_model).__name__,
    "training_timestamp_utc": datetime.utcnow().isoformat(),
    "target_column": TARGET_COLUMN,
    # "confidence_threshold": CONFIDENCE_THRESHOLD,
    "feature_engineering_file": "feature_engineering.py", # Указываем источник фичей
    "training_data_file": final_dataset_filename,
    "model_parameters": None, # final_model.get_params(),
    "feature_importance": None, # ToDo feature_importance.head(15).to_dict('records'),
    # "cross_validation_summary": results_summary.to_dict('records')
}

with open(metadata_path, 'w', encoding='utf-8') as f:
    # Используем indent=4 для красивого форматирования, ensure_ascii=False для кириллицы
    json.dump(metadata, f, indent=4, ensure_ascii=False)


# 2. Сохранение списка признаков
features_path = os.path.join(output_folder, f"{filename_base}_features.json")
features_list = list(X.columns) 
with open(features_path, 'w') as f:
    json.dump(features_list, f, indent=4)
    
print(f"Список признаков сохранен в: {features_path}")
print(f"Метаданные и результаты сохранены в: {metadata_path}")

print("\nВсе артефакты модели успешно сохранены!")
